+++ test/fixtures/many/b-pass-3.js
@@ -1,0 +1,7 @@
+console.log('b');
+
+setTimeout(() => {
+  console.log('boo');
+}, 3000);
+
+console.log('Arguments are', process.argv.slice(2));