+++ test/fixtures/many/c-fail-1.js
@@ -1,0 +1,10 @@
+console.log('c');
+
+setTimeout(() => {
+  console.log('crap');
+}, 3000);
+
+console.log('Typeof Array.prototype.values is', typeof ['defined'].values);
+console.log('Arguments are', process.argv.slice(2));
+
+this.is.undefined();