+++ test/fixtures/many/d.js
@@ -1,0 +1,7 @@
+console.log('d');
+
+setTimeout(() => {
+  console.log('damn');
+}, 3000);
+
+console.log(['defined'].values());